stbt: add a method `process_all_frames` available for the tests
This method forces the pipeline to process all the frames for the duration
of the calls.

This is to use when ad-hoc events are to be captured or when precision is
required.

Conflicts:
	stbt-run
	stbt.py

Completing process_all_frames() to get all the images from a button press to the frame match.
Still to fix/add:
1) Catch up with the live stream.
2) Process the images so as to get a measurement calculated from the images captured during the key press.

Added timestamp retrieve method, used by and renamed variable in case more handlers
need to be added.

Made get_timestamp public in stbt-run.

process_all_frames() now catches up with the live stream afters lowing
down to process all the frames.

However, there should be a way to do the same thing by substituting the
next line for something more efficient:
      self.restart_source_bin()

catch_up_live_stream() now has the code from restart_ource_bin() without
the underrun handler. This fixes the constant underruns and hanging that
it was causing between the first step and the second of the test.

get_buffers has been rnamed to get_last_frames for clearance, and it has
been set as public.

Added necessary changes for the gstreamer's queue attributes when
processing all frames. Now, the buffer doesn't get full and it is
possible to process correctly all the frames.

Code related to the pipeline configuration parameters has been grouped
into the new methods save_config() and restore_last_saved_config().