run-tests.sh: Added unit tests for process_all_frames().
Added three new unit tests to check integrity of `process_all_frames`. All
of them use a source with a timestamp overlapped on its frames, and set
its frame rate to 50fps.

1) test_match_timed_frame: waits for the specific match of a frame that
has a particular timestamp.

2) test_live_stream_caught_up_after_process_all_frames: while processing
all frames, it changes the videotestsrc pattern, waits for a specific match
with a timestamp and when it catches up with the live stream, it waits for
a match for that pattern.

3) test_match_consecutive_timed_frames_in_leaky_mode: tries to wait for
three matches corresponding to three consecutive frames, and expects it to
fail.