run-tests.sh: Review process_all_frames - Add a failing unit test to show process_all_frames is not working
This unit test is testing that process_all_frames makes it possible to
match an image appearing after an user action even if the frame does not
remain for a long time on the screen.

Unfortunately, the test is failing since it appears that after changing the
pattern of videotestsrc by pressing 18, the first frame processed by
stbt-templatematch is not the frame expected.
This might indicate that in the context of process_all_frames, some frames
are lost, unless that when changing the pattern of videotestsrc, the first
frames are skipped in case the test would have to be rewritten.

This also remove unused images and potentially flaky tests:
test_match_timed_frame
test_match_consecutive_timed_frames_in_leaky_mode